Create Okafor kosisochukwu marycynthia 2024704071
This Python program implements a grading system using `if` and `elif` statements. It takes a student's score as input and prints the corresponding grade based on the following scale:

- 70–100: Grade A
- 60–69: Grade B
- 50–59: Grade C
- 45–49: Grade D
- 40–44: Grade E
- 20–39: Grade F

For each valid score, the program also prints "YES" to confirm the grading was successful. If the score is outside the valid range, it prints "Invalid score" and "NO".

Created by: Okafor Kosisochukwu Marycynthia 
Reg. No:2024704071